[#50] feat: validation 추가

+++ src/main/java/com/modeunsa/shared/product/dto/ProductResponse.java
@@ -4,6 +4,7 @@ import com.modeunsa.boundedcontext.product.domain.ProductCategory;
 import com.modeunsa.boundedcontext.product.domain.ProductCurrency;
 import com.modeunsa.boundedcontext.product.domain.ProductStatus;
 import com.modeunsa.boundedcontext.product.domain.SaleStatus;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -18,8 +19,8 @@ public class ProductResponse {
   private String name;
   private ProductCategory category;
   private String description;
-  private int price;
-  private int salePrice;
+  private BigDecimal price;
+  private BigDecimal salePrice;
   private ProductCurrency currency;
   private ProductStatus productStatus;
   private SaleStatus saleStatus;
